requirements.txt
pandas
matplotlib
seaborn
re
sklearn.preprocessing
sklearn.model_selection
sklearn.ensemble
sklearn.metrics
sklearn.cluster
sklearn.decomposition
math